Migrating parsing functions to the file parsing.ts.
